Improved cartridge controller, service, exceptions

+++ src/main/java/com/andrewsavich/bajter/cartridgerefillservice/service/cartridge/CartridgeServiceImpl.java
@@ -1,5 +1,7 @@
 package com.andrewsavich.bajter.cartridgerefillservice.service.cartridge;
 
+import com.andrewsavich.bajter.cartridgerefillservice.exception.cartridge.CartridgeNotFoundException;
+import com.andrewsavich.bajter.cartridgerefillservice.exception.cartridge.CartridgeUniqueIdentifyException;
 import com.andrewsavich.bajter.cartridgerefillservice.model.cartridge.Cartridge;
 import com.andrewsavich.bajter.cartridgerefillservice.repository.CartridgeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,10 +10,11 @@ import org.springframework.stereotype.Service;
 import java.util.Comparator;
 import java.util.Date;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
-public class CartridgeServiceImpl implements CartridgeService{
+public class CartridgeServiceImpl implements CartridgeService {
 
     private final String START_UNIQUE_IDENTIFY_NUMBER = "B2500";
     private int lastGeneratedUniqueIdentifyNumber;
@@ -26,21 +29,35 @@ public class CartridgeServiceImpl implements CartridgeService{
 
     @Override
     public Cartridge getCartridgeById(Long id) {
-        return cartridgeRepository.findById(id).get();
+        Cartridge cartridge = cartridgeRepository.findById(id).orElse(null);
+
+        if (Objects.isNull(cartridge)) {
+            throw new CartridgeNotFoundException("Cartridge with id '" + id + "' not found");
+        }
+
+        return cartridge;
     }
 
     @Override
-    public Cartridge getCartridgeUniqueIdentify(String uniqueIdentify) {
-        return cartridgeRepository.findByUniqueIdentify(uniqueIdentify);
+    public Cartridge getCartridgeByUniqueIdentify(String uniqueIdentify) {
+        Cartridge cartridge = cartridgeRepository.findByUniqueIdentify(uniqueIdentify);
+
+        if (Objects.isNull(cartridge)) {
+            throw new CartridgeNotFoundException("Cartridge with uniqueIdentify '" + uniqueIdentify + "' not found");
+        }
+
+        return cartridge;
     }
 
     @Override
     public Cartridge saveCartridge(Cartridge cartridge) {
-        if(cartridge.getUniqueIdentify() == null){
-            cartridge.setUniqueIdentify(generateUniqueIdentify());
+
+        if (isExistCartridgeUniqIdentify(cartridge)) {
+            throw new CartridgeUniqueIdentifyException("Cartridge with unique identify ' " + cartridge.getUniqueIdentify() + "' exists");
         }
 
-        if(cartridge.getAddedDate() == null){
+        if (Objects.isNull(cartridge.getUniqueIdentify())) {
+            cartridge.setUniqueIdentify(generateUniqueIdentify());
             cartridge.setAddedDate(new Date());
         }
 
@@ -48,36 +65,34 @@ public class CartridgeServiceImpl implements CartridgeService{
     }
 
     @Override
-    public void deleteCartridge(Cartridge cartridge) {
+    public void deleteCartridgeById(Long id) {
+        Cartridge cartridge = cartridgeRepository.findById(id).orElse(null);
+
+        if (Objects.isNull(cartridge)) {
+            throw new CartridgeNotFoundException("Cartridge with id '" + id + "' not found");
+        }
+
         cartridgeRepository.delete(cartridge);
     }
 
-    @Override
-    public boolean isExistCartridgeUniqIdentify(Cartridge checkingCartridge) {
+
+    private boolean isExistCartridgeUniqIdentify(Cartridge checkingCartridge) {
         Cartridge existingCartridge = cartridgeRepository.findByUniqueIdentify(checkingCartridge.getUniqueIdentify());
 
-        if(checkingCartridge.getId() == null){
+        if (checkingCartridge.getId() == null) {
             return existingCartridge != null;
         }
 
-        if(existingCartridge == null){
+        if (existingCartridge == null) {
             return false;
-        }else{
+        } else {
             return existingCartridge.getId() != checkingCartridge.getId();
         }
 
     }
 
-    @Override
-    public void updateFields(Cartridge oldCartridge, Cartridge newCartridge) {
-        oldCartridge.setModel(newCartridge.getModel());
-        oldCartridge.setClient(newCartridge.getClient());
-        oldCartridge.setDescription(newCartridge.getDescription());
-    }
-
-    private String generateUniqueIdentify(){
-        List<Cartridge> sortedCartridges = cartridgeRepository.findAll().stream()
-                .sorted(Comparator.comparing(Cartridge::getUniqueIdentify)).collect(Collectors.toList());
+    private String generateUniqueIdentify() {
+        List<Cartridge> sortedCartridges = cartridgeRepository.findAll().stream().sorted(Comparator.comparing(Cartridge::getUniqueIdentify)).collect(Collectors.toList());
 
         if (sortedCartridges.isEmpty()) {
             return START_UNIQUE_IDENTIFY_NUMBER;
@@ -90,9 +105,7 @@ public class CartridgeServiceImpl implements CartridgeService{
         lastGeneratedUniqueIdentifyNumber = Integer.parseInt(lastUniqIdentifyStr.substring(1));
         ++lastGeneratedUniqueIdentifyNumber;
 
-        String uniqueIdentifyResult = "B" + String.valueOf(lastGeneratedUniqueIdentifyNumber);
-
-        return uniqueIdentifyResult;
+        return "B" + lastGeneratedUniqueIdentifyNumber;
     }
 
 }
